refactor(collection): moved collection to the support namespace - renamed Phalcon\Collection\Collection to Phalcon\Support\Collection
Refs #131
