+++ tests/unit/Support/Collection/Collection/ClearCest.php
