+++ tests/unit/Support/Collection/Collection/CountCest.php
