+++ tests/unit/Support/Collection/Collection/GetIteratorCest.php
