+++ tests/unit/Support/Collection/Collection/HasCest.php
