+++ tests/unit/Support/Collection/Collection/RemoveCest.php
