+++ tests/unit/Support/Collection/Collection/ToJsonCest.php
