+++ tests/unit/Support/Collection/ReadOnly/ClearCest.php
