+++ tests/unit/Support/Collection/ReadOnly/ConstructCest.php
