+++ tests/unit/Support/Collection/ReadOnly/GetIteratorCest.php
