+++ tests/unit/Support/Collection/ReadOnly/HasCest.php
