+++ tests/unit/Support/Collection/ReadOnly/JsonSerializeCest.php
