+++ tests/unit/Support/Collection/ReadOnly/RemoveCest.php
