+++ tests/unit/Support/Collection/ReadOnly/ToArrayCest.php
