+++ tests/unit/Support/Collection/ReadOnly/ToJsonCest.php
